Shift CustomizeDialog to datadialogs.py
This allows the plotview customizations of the Qt checkboxes, comboboxes, and pushbuttons to be used in datadialogs.py without causing a circular import.
